+++ tests/resources/test_opensearch_domain.py
@@ -1,0 +1,171 @@
+import pytest
+from pydantic import ValidationError
+
+from pycfmodel.model.resources.opensearch_domain import OpenSearchDomain
+
+
+@pytest.fixture()
+def valid_empty_opensearch_domain():
+    return OpenSearchDomain(
+        **{
+            "Type": "AWS::OpenSearchService::Domain",
+            "Properties": {},
+        }
+    )
+
+
+@pytest.fixture()
+def valid_opensearch_domain_from_aws_documentation_examples():
+    return OpenSearchDomain(
+        **{
+            "Type": "AWS::OpenSearchService::Domain",
+            "Properties": {
+                "AccessPolicies": {
+                    "Version": "2012-10-17",
+                    "Statement": [
+                        {
+                            "Effect": "Allow",
+                            "Principal": {"AWS": "arn:aws:iam::123456789012:user/opensearch-user"},
+                            "Action": "es:*",
+                            "Resource": "arn:aws:es:us-east-1:123456789012:domain/test/*",
+                        }
+                    ],
+                },
+                "AdvancedOptions": {
+                    "rest.action.multi.allow_explicit_index": True,
+                    "override_main_response_version": True,
+                },
+                "ClusterConfig": {
+                    "DedicatedMasterEnabled": True,
+                    "InstanceCount": "2",
+                    "ZoneAwarenessEnabled": True,
+                    "InstanceType": "m3.medium.search",
+                    "DedicatedMasterType": "m3.medium.search",
+                    "DedicatedMasterCount": "3",
+                },
+                "DomainName": "test",
+                "EBSOptions": {"EBSEnabled": True, "Iops": "0", "VolumeSize": "20", "VolumeType": "gp2"},
+                "EngineVersion": "OpenSearch_1.0",
+                "LogPublishingOptions": {
+                    "ES_APPLICATION_LOGS": {
+                        "CloudWatchLogsLogGroupArn": "arn:aws:logs:us-east-1:123456789012:log-group:/aws/opensearch/domains/opensearch-application-logs",
+                        "Enabled": True,
+                    },
+                    "SEARCH_SLOW_LOGS": {
+                        "CloudWatchLogsLogGroupArn": "arn:aws:logs:us-east-1:123456789012:log-group:/aws/opensearch/domains/opensearch-slow-logs",
+                        "Enabled": True,
+                    },
+                    "INDEX_SLOW_LOGS": {
+                        "CloudWatchLogsLogGroupArn": "arn:aws:logs:us-east-1:123456789012:log-group:/aws/opensearch/domains/opensearch-index-slow-logs",
+                        "Enabled": True,
+                    },
+                },
+            },
+        }
+    )
+
+
+def test_valid_empty_opensearch_domain_resource_can_be_built(valid_empty_opensearch_domain):
+    assert valid_empty_opensearch_domain.Type == "AWS::OpenSearchService::Domain"
+    assert valid_empty_opensearch_domain.Properties.AccessPolicies is None
+    assert valid_empty_opensearch_domain.Properties.AdvancedOptions is None
+    assert valid_empty_opensearch_domain.Properties.AdvancedSecurityOptions is None
+    assert valid_empty_opensearch_domain.Properties.ClusterConfig is None
+    assert valid_empty_opensearch_domain.Properties.CognitoOptions is None
+    assert valid_empty_opensearch_domain.Properties.DomainEndpointOptions is None
+    assert valid_empty_opensearch_domain.Properties.DomainName is None
+    assert valid_empty_opensearch_domain.Properties.EBSOptions is None
+    assert valid_empty_opensearch_domain.Properties.EncryptionAtRestOptions is None
+    assert valid_empty_opensearch_domain.Properties.EngineVersion is None
+    assert valid_empty_opensearch_domain.Properties.LogPublishingOptions is None
+    assert valid_empty_opensearch_domain.Properties.NodeToNodeEncryptionOptions is None
+    assert valid_empty_opensearch_domain.Properties.SnapshotOptions is None
+    assert valid_empty_opensearch_domain.Properties.Tags is None
+    assert valid_empty_opensearch_domain.Properties.VPCOptions is None
+
+
+def test_valid_opensearch_domain_from_aws_documentation_examples_resource_can_be_built(
+    valid_opensearch_domain_from_aws_documentation_examples,
+):
+    assert valid_opensearch_domain_from_aws_documentation_examples.Type == "AWS::OpenSearchService::Domain"
+    assert (
+        valid_opensearch_domain_from_aws_documentation_examples.Properties.AccessPolicies.Statement[0].Effect == "Allow"
+    )
+    assert (
+        valid_opensearch_domain_from_aws_documentation_examples.Properties.AccessPolicies.Statement[0].Action == "es:*"
+    )
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.AccessPolicies.Statement[0].Principal == {
+        "AWS": "arn:aws:iam::123456789012:user/opensearch-user"
+    }
+    assert (
+        valid_opensearch_domain_from_aws_documentation_examples.Properties.AccessPolicies.Statement[0].Resource
+        == "arn:aws:es:us-east-1:123456789012:domain/test/*"
+    )
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.AdvancedOptions == {
+        "rest.action.multi.allow_explicit_index": True,
+        "override_main_response_version": True,
+    }
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.AdvancedSecurityOptions is None
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.ClusterConfig == {
+        "DedicatedMasterEnabled": True,
+        "InstanceCount": "2",
+        "ZoneAwarenessEnabled": True,
+        "InstanceType": "m3.medium.search",
+        "DedicatedMasterType": "m3.medium.search",
+        "DedicatedMasterCount": "3",
+    }
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.CognitoOptions is None
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.DomainEndpointOptions is None
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.DomainName == "test"
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.EBSOptions == {
+        "EBSEnabled": True,
+        "Iops": "0",
+        "VolumeSize": "20",
+        "VolumeType": "gp2",
+    }
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.EngineVersion == "OpenSearch_1.0"
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.EncryptionAtRestOptions is None
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.LogPublishingOptions == {
+        "ES_APPLICATION_LOGS": {
+            "CloudWatchLogsLogGroupArn": "arn:aws:logs:us-east-1:123456789012:log-group:/aws/opensearch/domains/opensearch-application-logs",
+            "Enabled": True,
+        },
+        "SEARCH_SLOW_LOGS": {
+            "CloudWatchLogsLogGroupArn": "arn:aws:logs:us-east-1:123456789012:log-group:/aws/opensearch/domains/opensearch-slow-logs",
+            "Enabled": True,
+        },
+        "INDEX_SLOW_LOGS": {
+            "CloudWatchLogsLogGroupArn": "arn:aws:logs:us-east-1:123456789012:log-group:/aws/opensearch/domains/opensearch-index-slow-logs",
+            "Enabled": True,
+        },
+    }
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.NodeToNodeEncryptionOptions is None
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.SnapshotOptions is None
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.Tags is None
+    assert valid_opensearch_domain_from_aws_documentation_examples.Properties.VPCOptions is None
+
+
+def test_raise_error_if_invalid_fields_in_resource():
+    with pytest.raises(ValidationError) as exc_info:
+        OpenSearchDomain(
+            **{
+                "Type": "AWS::OpenSearchService::Domain",
+                "Properties": {
+                    "DomainName": [],
+                },
+            }
+        )
+
+    assert exc_info.value.errors() == [
+        {"loc": ("Properties", "DomainName"), "msg": "str type expected", "type": "type_error.str"},
+        {
+            "loc": ("Properties", "DomainName", "__root__"),
+            "msg": "FunctionDict should only have 1 key and be a function",
+            "type": "value_error",
+        },
+        {
+            "loc": ("Properties", "__root__"),
+            "msg": "FunctionDict should only have 1 key and be a function",
+            "type": "value_error",
+        },
+    ]